+++ packages/number-keyboard/index.js
@@ -1,0 +1,7 @@
+import NumberKeyboard from './src/main.vue'
+
+NumberKeyboard.install = Vue => {
+  Vue.component(NumberKeyboard.name, NumberKeyboard)
+}
+
+export default NumberKeyboard