+++ examples/docs/docs/changelog.md
@@ -19,10 +19,14 @@
   - 新增 `close-on-popstate` 属性，返回时自动关闭 (by [@yawuling](https://github.com/yawuling) )
 - MessageBox
   - 新增 `close-on-popstate` 属性，返回时自动关闭 (by [@yawuling](https://github.com/yawuling) )
+- NumberKeyboard
+  - 新增数字键盘组件 (by [@yawuling](https://github.com/yawuling) )
 - Picker
   - 新增 `close-on-popstate` 属性，返回时自动关闭 (by [@yawuling](https://github.com/yawuling) )
 - Popup
   - 新增 `close-on-popstate` 属性，返回时自动关闭 (by [@yawuling](https://github.com/yawuling) )
+- Search
+  - 新增 `type` 属性，支持设置 'number' 输入类型 (by [@yawuling](https://github.com/yawuling) )
 - SelectPicker
   - 新增 `close-on-popstate` 属性，返回时自动关闭 (by [@yawuling](https://github.com/yawuling) )